core/tryCatch.go
package core

import (
	"fmt"
)

// TryFunction corresponds to the try() part of a try / catch block
type TryFunction func()

// CatchFunction corresponds to the catch() part of a try / catch block
type CatchFunction func(error)

// TryCatch simulates a try/catch block using golang's recover() functionality
func TryCatch(try TryFunction, catch CatchFunction, catchFallbackMessage string) (ok bool) {
	defer func() {
		if r := recover(); r != nil {
			err, ok := r.(error)
			if !ok {
				err = fmt.Errorf("%s, caught panic: %v", catchFallbackMessage, r)
			}

			catch(err)
			ok = false
		}
	}()

	try()
	ok = true
	return
}
